moved memory to separate files
