+++ src/ojf13/include/mips_cpu_mem.h
@@ -1,0 +1,21 @@
+//
+//  mips_instr.h
+//  MIPSim
+//
+//  Created by Ollie Ford on 22/10/2014.
+//  Copyright (c) 2014 OJFord. All rights reserved.
+//
+
+struct mips_mem{
+public:
+	mips_mem(uint32_t, uint32_t);
+	~mips_mem();
+	
+	void read(uint8_t*, uint32_t, uint32_t) const;
+	void write(uint32_t, uint32_t, const uint8_t*);
+	
+private:
+	uint32_t	_size;
+	uint32_t	_bSize;
+	uint8_t*	_data;
+};
